model-test alteration, model-gen tweaks
Previous model was showing a tendency to overclass as ai. Tweaked model to include dropout, however it seems to be overfitting the data as the validation accuracy drops off after epoch 12 approx.

Model-test was altered to include a more detailed json of the tested file, category, scoring, off course closer to 0 indicates AI, closer to 1 indicates real

- included new matlab graph showing model accuracy.