Multiple different variations of model generation code. Revised new directory for stage 2
Played around with transfer learning models using resnet and VGG16.  Also with GPU as the compute for the model generation using tensorflow in WSL2 with CUDA.

These didnt actually return a higher accuracy on the unseen data unfortunately. Latest iteration of the revised stage 2 model is giving a fair but not very high accuracy. Classifying real images as real slightly over half the time but the AI detection element is quite fair, at around 75~% accuracy.

Continuing to play aorund with this in the hope of improving but wont be doing much more than this in the way of implemenetation instead finishing the report over the next few days.

I may do some experiments with data augmentation of the images to try emphasise the difference in diffusion generated images prior to feeding the images to the model.

NOTE: The transfer models exceeded the github repo size limit, these will be added in a separate directory.